created QuantumState and MeasurementOutcomes objects + adapted np backend execute_circuit
